Refactor: showProfie service function 위치 변경
