+++ src/main/java/com/kwcapstone/Controller/MainController.java
@@ -4,7 +4,9 @@ import com.kwcapstone.Common.Response.BaseResponse;
 import com.kwcapstone.Common.Response.SuccessStatus;
 import com.kwcapstone.Security.PrincipalDetails;
 import com.kwcapstone.Service.MainService;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
+import org.bson.types.ObjectId;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -58,5 +60,15 @@ public class MainController {
                                        @RequestParam(value = "tap", defaultValue = "entire") String tap,
                                        @RequestParam(value = "keyword") String keyword) {
         return BaseResponse.res(SuccessStatus.MAIN_SEARCH, mainService.searchProject(memberId, tap, keyword));
+    }
+
+
+    //프로필 띄우기
+    @Operation(summary = "프로필 띄우기")
+    @GetMapping("/profile")
+    public BaseResponse showProfile(@AuthenticationPrincipal PrincipalDetails principalDetails){
+        ObjectId memberId = principalDetails.getId();
+
+
     }
 }
